amp up kP to pass test for now

+++ src/main/java/org/team4159/lib/hardware/Gearing.java
@@ -3,11 +3,11 @@ package org.team4159.lib.hardware;
 import org.team4159.lib.math.MathUtil;
 
 public class Gearing {
-  public int COUNTS_PER_REV;
+  public double COUNTS_PER_REV;
   public double COUNTS_PER_DEGREE, COUNTS_PER_RADIAN;
 
   public Gearing(double gear_ratio, int encoder_cpr) {
-    COUNTS_PER_REV = (int) (gear_ratio * encoder_cpr);
+    COUNTS_PER_REV = gear_ratio * encoder_cpr;
     COUNTS_PER_DEGREE = COUNTS_PER_REV / 360.0;
     COUNTS_PER_RADIAN = COUNTS_PER_REV / MathUtil.kTau;
   }
